requirements/base.txt
django==1.11.5
django-braces>=1.12.0
django-extensions>=2.0.6
django-guardian>=1.4.9
django-reversion>=2.0.9
django-widget-tweaks>=1.4.2
djangorestframework>=3.8.2
djangorestframework-filters>=0.10.2
social-auth-app-django>=1.2.0
factory-boy>=2.9.2
numpy>=1.12.1
celery>=4.1.1
flower>=0.9.2
psycopg2-binary
pypandoc>=1.4
selenium>=3.4.3
Faker>=0.7.18
gunicorn
whitenoise
dj-database-url
git+https://github.com/afrubin/idutils.git@bio
hg+http://bitbucket.org/afrubin/metapub@html_citation
git+https://github.com/FowlerLab/hgvs-patterns.git
